Read data as np array, not list. Made time linear.

+++ Parser.py
@@ -1,5 +1,4 @@
 import os
-import re
 import numpy as np
 import shutil
 import zipfile
@@ -11,48 +10,44 @@ def parseFilesInFolder(filename: str,
                         ):
 
     z = zipfile.ZipFile(zipFilePath)
-    in_file = z.open(filename,'r') # open file in zip file
+    in_file = z.open(filename, mode='r') # open file in zip file
     in_data = in_file.read().splitlines()[4:] # read in data from file in zip file
-    outfile = open(out_dir+''.join(filename[filename.find('STAR'):filename.find('_',filename.find('STAR'))]) + '.txt', 'a') # names output file as STARLINK-####.txt
-    
-    # Only get the data for the date specified in the file name
-    current_time = float(in_data[0].split()[0])
-    tf = current_time+1000000.726
-    idx = 0
-    while current_time <= tf:
-        toks = in_data[idx].split()
-        # reads in position and velocity values from data and converts into m and m/s respectively
-        pv = np.array(toks[1:],dtype = float)*1000
-        # calculates  angular momentum of the data
-        h = np.cross(pv[0:3], pv[3:6])
-        # Stores position, velocity, and angular momentum as a string
-        pvh = np.array2string(np.concatenate((pv,h)))
-        # cleans up the string for storage
-        pvh = re.sub('\[','',pvh)
-        pvh = re.sub('\]','',pvh)
-        pvh = re.sub('\n','',pvh)
-        # adds converts time in to string and cleans up the string
-        time = str(toks[0])
-        time = re.sub('^b\'','', time)
-        time = re.sub('\'$', ' ', time)
-        # writes to the outputfile
-        outfile.write(time)
-        outfile.write(pvh)
-        outfile.write('\n')
 
-        idx = idx + 20
-        current_time = float(toks[0])
+    numofmin = 5 
+    in_data = np.loadtxt(in_data)[::4*numofmin] # grabs data once every numofmin
 
-    in_file.close()
-    outfile.close()
+    # names output directory and filename = STARLINK-####.txt
+    outfile_name = out_dir + ''.join(filename[filename.find('STAR'):filename.find('_', filename.find('STAR'))]) + '.txt'
+
+    # selects the first 24 hours of indices
+    ind = np.arange(0, int(60/numofmin*24))
+    juliantime = in_data[ind, 0:1] 
+    pv = in_data[ind, 1:] # position and velocity
+    h = np.cross(pv[:, 0:3], pv[:, 3:6]) # calculates angular momentum which is r cross v
+
+    out_data = np.concatenate((juliantime, pv, h), axis=1)
+
+    # Checks if previously saved data overlaps with the start time of current data.
+    if os.path.isfile(outfile_name) == True:
+        # deletes the overlaps in the previous data
+        last_data = np.loadtxt(outfile_name)
+        first_time = in_data[0, 0]
+        repeats = np.argwhere(last_data[:, 0] >= first_time)
+        last_data = np.delete(last_data, repeats, axis=0)
+        # stacks the new data under the old data without any overlaps
+        out_data = np.concatenate((last_data, out_data), axis=0)
+
+    np.savetxt(outfile_name, out_data)
 
 if (__name__ == "__main__"):
     rootdir = '/scratch/08550/qz3485/SpaceX_ML'
     out_dir = '/home1/08550/qz3485/workspace/output_data/'
 
     # deletes the output directory if it exists already
-    shutil.rmtree(out_dir, ignore_errors=True)
-    os.makedirs('output_data', exist_ok = True)
+    deletedir = input("Want to delete output data? (1 = yes, 0 = no) ")
+    if deletedir == "1":
+        shutil.rmtree(out_dir, ignore_errors=True)
+        os.makedirs(out_dir, exist_ok=True)
     
     multiprocessing.set_start_method("spawn")
     # loops through each zipfile in rootdir
@@ -61,8 +56,8 @@ if (__name__ == "__main__"):
         files.sort()
 
         for file in files: 
-            zipFilePath = os.path.join(subdir,file)
-            print('Extracting from '+ zipFilePath)
+            zipFilePath = os.path.join(subdir, file)
+            print('Extracting from ' + zipFilePath)
             with zipfile.ZipFile(zipFilePath) as z: # go through zip files    
                 filepath_list = []
                 for filename in z.namelist(): # parse files in the .zip file
